Changed prior on log_theta
The uniform prior gave a log-prob of -Inf if the values fell outside the
bounds. This caused divergent transitions for samples with very low
concentrations. Changed to a normal, fairly vague. A rapid comparison of
the results from the two models suggests little to no effect of the
normal prior.