Model was re-written
The model was re-written from scratch.
An issue with the previous structure was that every well was considered "independent" within a sample. In reality, we perform each dilution once, and run 2-3 technical replicates. The previous model had trouble estimating the two errors (in x and y) because both could fulfill the same function. The new models are structured so that every dilution uses the same estimate in X, which allows the error in y to be properly estimated.